created separate file for Nav Bar
